virtual-tryon-backend/requirements.txt
# requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# AI/ML 라이브러리
torch==2.0.1
torchvision==0.15.2
opencv-python==4.8.1.78
mediapipe==0.10.7
numpy==1.24.3
pillow==10.0.1
scikit-image==0.21.0

# 이미지 처리
albumentations==1.3.1
imageio==2.31.6

# 성능 최적화
redis==5.0.1
celery==5.3.4

# 로깅 및 모니터링
loguru==0.7.2
prometheus-client==0.19.0